RF: fix vec_val_vect logic, generalize for shape
There was no need to enforce the input vec array to have the same length
for dimensions [-2:], I got confused, possibly because I had had a
couple of glasses of wine when I was writing this.  Remove that
requirement and test for other values of shape[-2], shape[-1].

The routine could easily be generalized to accept another input array
for the vec.T part.